LPS-77456 Inline WikiCacheHelper, it has a circular dependency on WikiPageLocalService.
