+++ modules/apps/collaboration/wiki/wiki-service/src/main/java/com/liferay/wiki/service/impl/WikiNodeLocalServiceImpl.java
@@ -18,6 +18,8 @@ import com.liferay.document.library.kernel.model.DLFolderConstants;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
 import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
 import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.cache.MultiVMPool;
+import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.log.Log;
@@ -52,8 +54,8 @@ import com.liferay.wiki.importer.WikiImporter;
 import com.liferay.wiki.internal.util.WikiCacheThreadLocal;
 import com.liferay.wiki.model.WikiNode;
 import com.liferay.wiki.model.WikiPage;
+import com.liferay.wiki.model.WikiPageDisplay;
 import com.liferay.wiki.service.base.WikiNodeLocalServiceBaseImpl;
-import com.liferay.wiki.util.WikiCacheHelper;
 
 import java.io.InputStream;
 
@@ -200,6 +202,9 @@ public class WikiNodeLocalServiceImpl extends WikiNodeLocalServiceBaseImpl {
 		_wikiImporterServiceTrackerMap =
 			ServiceTrackerMapFactory.openSingleValueMap(
 				bundle.getBundleContext(), WikiImporter.class, "importer");
+
+		_portalCache = multiVMPool.getPortalCache(
+			WikiPageDisplay.class.getName());
 	}
 
 	@Override
@@ -228,7 +233,7 @@ public class WikiNodeLocalServiceImpl extends WikiNodeLocalServiceBaseImpl {
 		finally {
 			WikiCacheThreadLocal.setClearCache(clearCache);
 
-			wikiCacheHelper.clearCache();
+			_portalCache.removeAll();
 		}
 
 		// Node
@@ -408,7 +413,7 @@ public class WikiNodeLocalServiceImpl extends WikiNodeLocalServiceBaseImpl {
 			NotificationThreadLocal.setEnabled(notificationsEnabled);
 			WikiCacheThreadLocal.setClearCache(clearCache);
 
-			wikiCacheHelper.clearCache();
+			_portalCache.removeAll();
 		}
 	}
 
@@ -591,7 +596,7 @@ public class WikiNodeLocalServiceImpl extends WikiNodeLocalServiceBaseImpl {
 		finally {
 			WikiCacheThreadLocal.setClearCache(clearCache);
 
-			wikiCacheHelper.clearCache();
+			_portalCache.removeAll();
 		}
 	}
 
@@ -631,6 +636,9 @@ public class WikiNodeLocalServiceImpl extends WikiNodeLocalServiceBaseImpl {
 		validate(0, groupId, name);
 	}
 
+	@ServiceReference(type = MultiVMPool.class)
+	protected MultiVMPool multiVMPool;
+
 	@ServiceReference(type = SubscriptionLocalService.class)
 	protected SubscriptionLocalService subscriptionLocalService;
 
@@ -640,15 +648,13 @@ public class WikiNodeLocalServiceImpl extends WikiNodeLocalServiceBaseImpl {
 	@ServiceReference(type = TrashHelper.class)
 	protected TrashHelper trashHelper;
 
-	@ServiceReference(type = WikiCacheHelper.class)
-	protected WikiCacheHelper wikiCacheHelper;
-
 	@ServiceReference(type = WikiGroupServiceConfiguration.class)
 	protected WikiGroupServiceConfiguration wikiGroupServiceConfiguration;
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		WikiNodeLocalServiceImpl.class);
 
+	private PortalCache<?, ?> _portalCache;
 	private ServiceTrackerMap<String, WikiImporter>
 		_wikiImporterServiceTrackerMap;
